requirements.txt
streamlit
openai
pdfplumber
pandas
openpyxl
python-dotenv
